Share serialization tests.
